+++ src/main/java/org/rgt/libraryhub/repository/PatronRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PatronRepository extends JpaRepository<Patron, Integer> {}
+public interface PatronRepository extends JpaRepository<Patron, Long> {}
